+++ opentelemetry-sdk/tests/metrics/test_metrics.py
@@ -314,6 +314,138 @@ class TestValueRecorder(unittest.TestCase):
         )
 
 
+class TestSumObserver(unittest.TestCase):
+    def test_observe(self):
+        meter = metrics.MeterProvider().get_meter(__name__)
+        observer = metrics.SumObserver(
+            None, "name", "desc", "unit", int, meter, ("key",), True
+        )
+        labels = {"key": "value"}
+        key_labels = tuple(sorted(labels.items()))
+        values = (37, 42, 60, 100)
+        for val in values:
+            observer.observe(val, labels)
+
+        self.assertEqual(observer.aggregators[key_labels].current, values[-1])
+
+    def test_observe_disabled(self):
+        meter = metrics.MeterProvider().get_meter(__name__)
+        observer = metrics.SumObserver(
+            None, "name", "desc", "unit", int, meter, ("key",), False
+        )
+        labels = {"key": "value"}
+        observer.observe(37, labels)
+        self.assertEqual(len(observer.aggregators), 0)
+
+    @mock.patch("opentelemetry.sdk.metrics.logger")
+    def test_observe_incorrect_type(self, logger_mock):
+        meter = metrics.MeterProvider().get_meter(__name__)
+        observer = metrics.SumObserver(
+            None, "name", "desc", "unit", int, meter, ("key",), True
+        )
+        labels = {"key": "value"}
+        observer.observe(37.0, labels)
+        self.assertEqual(len(observer.aggregators), 0)
+        self.assertTrue(logger_mock.warning.called)
+
+    @mock.patch("opentelemetry.sdk.metrics.logger")
+    def test_observe_non_decreasing_error(self, logger_mock):
+        meter = metrics.MeterProvider().get_meter(__name__)
+        observer = metrics.SumObserver(
+            None, "name", "desc", "unit", int, meter, ("key",), True
+        )
+        labels = {"key": "value"}
+        observer.observe(37, labels)
+        observer.observe(14, labels)
+        self.assertEqual(len(observer.aggregators), 1)
+        self.assertTrue(logger_mock.warning.called)
+
+    def test_run(self):
+        meter = metrics.MeterProvider().get_meter(__name__)
+
+        callback = mock.Mock()
+        observer = metrics.SumObserver(
+            callback, "name", "desc", "unit", int, meter, (), True
+        )
+
+        self.assertTrue(observer.run())
+        callback.assert_called_once_with(observer)
+
+    @mock.patch("opentelemetry.sdk.metrics.logger")
+    def test_run_exception(self, logger_mock):
+        meter = metrics.MeterProvider().get_meter(__name__)
+
+        callback = mock.Mock()
+        callback.side_effect = Exception("We have a problem!")
+
+        observer = metrics.SumObserver(
+            callback, "name", "desc", "unit", int, meter, (), True
+        )
+
+        self.assertFalse(observer.run())
+        self.assertTrue(logger_mock.warning.called)
+
+
+class TestUpDownSumObserver(unittest.TestCase):
+    def test_observe(self):
+        meter = metrics.MeterProvider().get_meter(__name__)
+        observer = metrics.UpDownSumObserver(
+            None, "name", "desc", "unit", int, meter, ("key",), True
+        )
+        labels = {"key": "value"}
+        key_labels = tuple(sorted(labels.items()))
+        values = (37, 42, 14, 30)
+        for val in values:
+            observer.observe(val, labels)
+
+        self.assertEqual(observer.aggregators[key_labels].current, values[-1])
+
+    def test_observe_disabled(self):
+        meter = metrics.MeterProvider().get_meter(__name__)
+        observer = metrics.UpDownSumObserver(
+            None, "name", "desc", "unit", int, meter, ("key",), False
+        )
+        labels = {"key": "value"}
+        observer.observe(37, labels)
+        self.assertEqual(len(observer.aggregators), 0)
+
+    @mock.patch("opentelemetry.sdk.metrics.logger")
+    def test_observe_incorrect_type(self, logger_mock):
+        meter = metrics.MeterProvider().get_meter(__name__)
+        observer = metrics.UpDownSumObserver(
+            None, "name", "desc", "unit", int, meter, ("key",), True
+        )
+        labels = {"key": "value"}
+        observer.observe(37.0, labels)
+        self.assertEqual(len(observer.aggregators), 0)
+        self.assertTrue(logger_mock.warning.called)
+
+    def test_run(self):
+        meter = metrics.MeterProvider().get_meter(__name__)
+
+        callback = mock.Mock()
+        observer = metrics.UpDownSumObserver(
+            callback, "name", "desc", "unit", int, meter, (), True
+        )
+
+        self.assertTrue(observer.run())
+        callback.assert_called_once_with(observer)
+
+    @mock.patch("opentelemetry.sdk.metrics.logger")
+    def test_run_exception(self, logger_mock):
+        meter = metrics.MeterProvider().get_meter(__name__)
+
+        callback = mock.Mock()
+        callback.side_effect = Exception("We have a problem!")
+
+        observer = metrics.UpDownSumObserver(
+            callback, "name", "desc", "unit", int, meter, (), True
+        )
+
+        self.assertFalse(observer.run())
+        self.assertTrue(logger_mock.warning.called)
+
+
 class TestValueObserver(unittest.TestCase):
     def test_observe(self):
         meter = metrics.MeterProvider().get_meter(__name__)