+++ opentelemetry-sdk/CHANGELOG.md
@@ -14,6 +14,8 @@
   ([#764](https://github.com/open-telemetry/opentelemetry-python/pull/764))
 - Add start_pipeline to MeterProvider
   ([#764](https://github.com/open-telemetry/opentelemetry-python/pull/764))
+- Add SumObserver, UpDownSumObserver and LastValueAggregator in metrics
+  ([#789](https://github.com/open-telemetry/opentelemetry-python/pull/789))
 
 ## 0.8b0